added missing functionality

+++ src/app/components/athletes/athlete-list/athlete-list.component.ts
@@ -3,6 +3,7 @@ import {Athlete} from '../../../api/api-interfaces';
 import {Observable} from 'rxjs';
 import {AthleteService} from '../service/athlete.service';
 import {getYearsFrom} from '../../../util/date-utils';
+import {AuthService} from '../../auth/service/auth.service';
 
 @Component({
   selector: 'app-athletes-list',
@@ -10,10 +11,11 @@ import {getYearsFrom} from '../../../util/date-utils';
 })
 export class AthleteListComponent {
 
-  headerNames = ['Imię', 'Nazwisko', 'Email', 'Numer telefonu', 'Wiek', 'Akcje'];
+  headerNames = ['Imię', 'Nazwisko', 'Email', 'Numer telefonu', 'Aktualny klub', 'Wiek', 'Akcje'];
   athletes$: Observable<Athlete[]>;
 
-  constructor(private service: AthleteService) {
+  constructor(private service: AthleteService,
+              private authService: AuthService) {
     this.athletes$ = service.getAthletes();
   }
 
@@ -22,7 +24,18 @@ export class AthleteListComponent {
   }
 
   deleteRecord(id: number) {
+    if (!this.authService.loggedIn()) {
+      return;
+    }
+
+    const confirmed = confirm('Na pewno chcesz usunąć rekord?');
+
+    if (!confirmed) {
+      return;
+    }
+
     this.service.deleteAthlete(id).subscribe(res => {
+      alert('Usunięto sportowca');
       window.location.reload();
     }, err => {
       console.log(err);
