+++ src/app/components/managers/manager-list/manager-list.component.ts
@@ -3,6 +3,7 @@ import {Observable} from 'rxjs';
 import {Manager} from '../../../api/api-interfaces';
 import {getYearsFrom} from '../../../util/date-utils';
 import {ManagerService} from '../service/manager.service';
+import {AuthService} from '../../auth/service/auth.service';
 
 @Component({
   selector: 'app-manager-list',
@@ -13,7 +14,8 @@ export class ManagerListComponent {
   headerNames = ['Imię', 'Nazwisko', 'Email', 'Numer telefonu', 'Staż (w latach)', 'Akcje'];
   managers$: Observable<Manager[]>;
 
-  constructor(private service: ManagerService) {
+  constructor(private service: ManagerService,
+              private authService: AuthService) {
     this.managers$ = service.getManagers();
   }
 
@@ -22,7 +24,18 @@ export class ManagerListComponent {
   }
 
   deleteRecord(id: number) {
+    if (!this.authService.loggedIn()) {
+      return;
+    }
+
+    const confirmed = confirm('Na pewno chcesz usunąć rekord?');
+
+    if (!confirmed) {
+      return;
+    }
+
     this.service.deleteManager(id).subscribe(res => {
+      alert('Usunięto menadżera');
       window.location.reload();
     }, err => {
       console.log(err);
