+++ src/app/components/transfers/transfer-list/transfer-list.component.ts
@@ -3,6 +3,7 @@ import {TransferService} from '../service/transfer.service';
 import {Observable} from 'rxjs';
 import {Transfer} from '../../../api/api-interfaces';
 import {getSimpleDateString} from '../../../util/date-utils';
+import {AuthService} from '../../auth/service/auth.service';
 
 
 @Component({
@@ -15,8 +16,11 @@ export class TransferListComponent {
     'Prowizja (%)', 'Wypożyczenie?', 'Kontrakt start', 'Kontrakt koniec', 'Akcje'];
   transfers$: Observable<Transfer[]>;
 
-  constructor(private service: TransferService) {
-    this.transfers$ = service.getTransfers();
+  constructor(private service: TransferService,
+              private authService: AuthService) {
+    if (authService.loggedIn()) {
+      this.transfers$ = service.getTransfers();
+    }
   }
 
   getSimpleDate(inputDate: string): string {
@@ -24,7 +28,14 @@ export class TransferListComponent {
   }
 
   deleteRecord(id: number) {
+    const confirmed = confirm('Na pewno chcesz usunąć rekord?');
+
+    if (!confirmed) {
+      return;
+    }
+
     this.service.deleteTransfer(id).subscribe(res => {
+      alert('Usunięto transfer');
       window.location.reload();
     }, err => {
       console.log(err);
